商品管理改造-商品删除

+++ dkd-manage/src/main/java/com/dkd/manage/service/impl/SkuServiceImpl.java
@@ -2,6 +2,7 @@ package com.dkd.manage.service.impl;
 
 import java.util.List;
 import com.dkd.common.utils.DateUtils;
+import com.dkd.manage.service.IChannelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.dkd.manage.mapper.SkuMapper;
@@ -19,6 +20,8 @@ public class SkuServiceImpl implements ISkuService
 {
     @Autowired
     private SkuMapper skuMapper;
+    @Autowired
+    private IChannelService channelService;
 
     /**
      * 查询商品管理
@@ -79,6 +82,12 @@ public class SkuServiceImpl implements ISkuService
     @Override
     public int deleteSkuBySkuIds(Long[] skuIds)
     {
+        //1. 判断商品id是否有关联货道
+        int count = channelService.countChannelBySkuIds(skuIds);
+        if(count > 0){
+            throw new RuntimeException("此商品被货道关联，无法删除");
+        }
+        //2. 没有关联货道可以直接删除
         return skuMapper.deleteSkuBySkuIds(skuIds);
     }
